LayoutTests/fast/multicol/margin-bottom-and-break-after.html

<!DOCTYPE html>
<p>A block that follows a block with -webkit-column-break-after:always; should be positioned at the
    *top* of the *next* column. Margins may need to be ignored to achieve this.</p>
<div id="mc" style="-webkit-column-count:2; column-fill:auto; width:100px; height:200px;">
    <div style="height:150px; margin-bottom:150px; -webkit-column-break-after:always;"></div>
    <div id="child" style="height:100px;"></div>
</div>
<div id="console"></div>
<script src="../../resources/js-test.js"></script>
<script>
    var mc = document.getElementById('mc');
    var child = document.getElementById('child');
    shouldBe("child.offsetLeft + child.offsetWidth", "mc.offsetLeft + mc.offsetWidth");
    shouldBe("child.offsetTop", "mc.offsetTop");
</script>
